% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_raw_data.r
\name{compare_raw_data}
\alias{compare_raw_data}
\title{Compare STI ORA}
\usage{
compare_raw_data(ref_collection = "STI")
}
\arguments{
\item{Str}{a character vector with UTF-8 encoding errors.}
}
\description{
create NEW, REV, DEL files at ref_area, freq level
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{

# to do

}
\seealso{
\code{\link{getCodelist}} \code{\link{getDataStructure}}
}
\author{
ILO bescond
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}
