% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.r
\name{DT}
\alias{DT}
\title{datatable custum for ilostat need}
\usage{
DT(X, apps = FALSE)
}
\arguments{
\item{X}{dataframe to view.}

\item{apps, }{option when embed in an apps, default FALSE.}
}
\description{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\author{
ILO bescond
}
