% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataOracle.r
\name{getDataOracle}
\alias{getDataOracle}
\title{Query Data ILOSTAT}
\usage{
getDataOracle(data_query, con)
}
\arguments{
\item{data_query}{SQL character string for select in table Data and Note, faster if split in 2: QTABLE and VALUE, check example.}

\item{Connect}{list of 4 characters string to initialiyse JDBC connection: list(
drv = JDBCDriver , a DBI driver connection, 
dns = 'jdbc_oracle_thin_client//my_localhost:my_port/my_service_name', 
user = 'my_user_name', 
pwd = 'my_password')}
}
\description{
Query data from ILOSTAT Oracle database
}
\details{
Helper function to efficiently query data from ILOSTAT Oracle database.
}
\examples{
################################## 
 head(res)					
}
\author{
ILO bescond
}
\keyword{DBI}
\keyword{ILO,}
\keyword{RJDBC,}
