% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataOracle.r
\name{getDataOracle_NEW}
\alias{getDataOracle_NEW}
\title{Query Data ILOSTAT}
\usage{
getDataOracle_NEW(data_query, extra = FALSE, Save = FALSE, mv = FALSE,
  con)
}
\arguments{
\item{data_query}{SQL character string for select in table Data and Note, faster if split in 2: QTABLE and VALUE, check example.}

\item{extra}{to get extra administrative info, default equal FALSE}

\item{Connect}{list of 4 characters string to initialiyse JDBC connection: list(
drv = JDBCDriver , a DBI driver connection, 
dns = 'jdbc_oracle_thin_client//my_localhost:my_port/my_service_name', 
user = 'my_user_name', 
pwd = 'my_password')}

\item{Codelist}{list tables from the internal ilostat codelist}
}
\description{
Query data from ILOSTAT Oracle database
}
\details{
Helper function to efficiently query data from ILOSTAT Oracle database.
}
\examples{
################################## 

WHERE <- c(	QTABLE = "where Collection_Code in ('STI') and 
							Indicator_Code in ('EMP_TEMP_SEX_ECO_NB','EES_TEES_SEX_ECO_NB','HOW_TEMP_SEX_ECO_NB','HOW_XEES_SEX_ECO_NB') and 
							Classif1_Version_Code in ('ECO_SECTOR','ECO_AGGREGATE') and
							Year > 2001",
			VALUE 	= "			Classif1_Code in ('ECO_SECTOR_NAG','ECO_AGGREGATE_MAN') and 
							Sex_Code in ('SEX_T')")
res <- getDataOracle(WHERE, Connect = initJ, Codelist = CODE_ORA) 
 head(res)					
}
\author{
ILO bescond
}
\keyword{DBI}
\keyword{ILO,}
\keyword{RJDBC,}
