% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMetaOracle.r
\name{LoadMetaOracle}
\alias{LoadMetaOracle}
\title{ilostat codelist definition}
\usage{
LoadMetaOracle(con, wd, wdAriane)
}
\description{
ilostat codelist definition
}
\author{
ILO bescond
}
\keyword{ILO}
