% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroTestClassifVersion.r
\name{MicroTestClassifVersion}
\alias{MicroTestClassifVersion}
\title{test data with ilostat framework}
\usage{
MicroTestClassifVersion(df, Collection, resume = FALSE)
}
\arguments{
\item{df, }{dataframe to reduce according to the ILO collection configuration.}

\item{Collection, }{Collection configuration to test.}
}
\description{
Helper function to efficiently load microdataset.
}
\author{
ILO bescond
}
\keyword{ILO,}
\keyword{R}
