% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_Label_notesJ.r
\name{My_Label_notesJ}
\alias{My_Label_notesJ}
\title{map code to label at factor levels efficiently in en, fr or sp language}
\usage{
My_Label_notesJ(col_note, SEP = " | ", Lang = "en")
}
\arguments{
\item{col_note}{reference columns code contains 0, 1 or multiple code like C2:158 | C3:485}

\item{SEP}{separator between multiple code, default " | "}

\item{Lang}{language, default = en, fr, sp also available}
}
\description{
map code to label at factor levels efficiently in en, fr or sp language
}
\examples{
################################# use to identify a proper	'DSD'		 

X \%>\%	mutate(	Notes_Source_Label = 	mapvalues(Notes_Source_Code,	from = levels(as.factor(Notes_Source_Code)), 
																		to = My_Label_notesJ(levels(as.factor(Notes_Source_Code)),SEP = "_", Lang = Lang)))

}
\author{
ILO bescond
}
\keyword{ILO}
