% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_Resort_NotesJ.r
\name{My_Resort_NotesJ}
\alias{My_Resort_NotesJ}
\title{resort code efficiently inside one columns in en, fr or sp language}
\usage{
My_Resort_NotesJ(col_note, SEP = " | ")
}
\arguments{
\item{col_note}{reference columns code contains 0, 1 or multiple code like C2:158 | C3:485}

\item{SEP}{separator between multiple code, default " | "}
}
\description{
resort code efficiently inside one columns in en, fr or sp language
}
\examples{
################################# use to identify a proper	'DSD'		 

X \%>\%	mutate(	Notes_Indicator_Code 	= mapvalues(Notes_Indicator_Code,	from = levels(as.factor(Notes_Indicator_Code)), 
																			to = My_Resort_NotesJ(levels(as.factor(Notes_Indicator_Code)),SEP = "_"), warn_missing = FALSE))

}
\author{
ILO bescond
}
\keyword{ILO}
