% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_Transform_notesJ.r
\name{My_Transform_notesJ}
\alias{My_Transform_notesJ}
\title{transform code efficiently inside one columns in en, fr or sp language used for csv upload format}
\usage{
My_Transform_notesJ(col_note, SEP = " | ")
}
\arguments{
\item{col_note}{reference columns code contains 0, 1 or multiple code like C2:158 | C3:485, will be transform as #485$NA_#485$NA}

\item{SEP}{separator between multiple code, default " | "}
}
\description{
transform code efficiently inside one columns in en, fr or sp language used for csv upload format
}
\examples{
################################# use to identify a proper	'DSD'		 

X \%>\%	mutate(	Notes_Source_Code = 	plyr:::mapvalues(Notes_Source_Code,	from = levels(as.factor(Notes_Source_Code)), 
																		to = My_Transform_notesJ(levels(as.factor(Notes_Source_Code)),SEP = "_")))

}
\author{
ILO bescond
}
\keyword{ILO}
