% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyDataST.r
\name{plyDataST}
\alias{plyDataST}
\title{ilo short term data processing}
\usage{
plyDataST(Title, ReadMe, mywd)
}
\arguments{
\item{Title}{character string of the country alpha iso code 3.}

\item{ReadMe}{file containing path of each ST original files.}

\item{mywd}{system work directory.}
}
\description{
ilo short term data processing
}
\author{
ILO bescond
}
\keyword{ILO}
