% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ilo.r
\name{prepare_segment_ilo}
\alias{prepare_segment_ilo}
\title{Prepare ilo  packages}
\usage{
prepare_segment_ilo(FREQ, DBNAme, query, cou)
}
\arguments{
\item{X}{data frame.}

\item{DELETE}{in deletion format, default is FALSE.}

\item{LABEL}{add label to the notes, default is FALSE.}
}
\description{
fix utf-8 encoding bug
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\author{
ILO bescond
}
\keyword{ILO,}
\keyword{R}
