% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareDownload.r
\name{PrepareDownload}
\alias{PrepareDownload}
\title{Prepare dataframe in csv download format}
\usage{
PrepareDownload(X, Lang = "en")
}
\arguments{
\item{X}{data frame.}

\item{Lang}{default "en" English, could be "fr", and "sp".}

\item{LABEL}{add label to the notes, default is FALSE.}
}
\description{
fix utf-8 encoding bug
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\author{
ILO bescond
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}
