% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortDataOracle.r
\name{sortDataOracle}
\alias{sortDataOracle}
\title{sort ilo data correctly}
\usage{
sortDataOracle(X, Lang = "EN")
}
\arguments{
\item{X}{data frame.}

\item{Lang}{language, default = EN, SP, FR also available}
}
\description{
sort ilo data correctly
}
\author{
ILO bescond
}
\keyword{ILO}
