% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyDataXQ.r
\name{XQ_Meta}
\alias{XQ_Meta}
\title{ilo short term data processing}
\usage{
XQ_Meta()
}
\arguments{
\item{Title}{character string of the country alpha iso code 3.}

\item{ReadMe}{file containing path of each ST original files.}
}
\description{
ilo short term data processing
}
\author{
ILO bescond
}
\keyword{ILO}
