% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_raw_data.r
\name{compare_raw_data}
\alias{compare_raw_data}
\title{Compare STI ORA}
\usage{
compare_raw_data(ref_collection = "STI", ref_ref_area = NULL,
  ref_freq_code = c("Q", "M"))
}
\arguments{
\item{ref_collection}{select your collection.}

\item{ref_area}{reduce to selected ref_area.}
}
\description{
create NEW, REV, DEL files at ref_area, freq level
}
\details{
Helper function to efficiently query data from ILOSTAT SDMX API.
}
\examples{

# to do

}
\author{
ILO bescond
}
\keyword{ILO,}
\keyword{R}
\keyword{SDMX,}
